+++ tests/unit/EnvironmentTest.php
@@ -155,13 +155,13 @@ class EnvironmentTest extends \PHPUnit_Framework_TestCase
             'BlockQuote'     => 'League\CommonMark\Block\Element\BlockQuote',
             'Document'       => 'League\CommonMark\Block\Element\Document',
             'FencedCode'     => 'League\CommonMark\Block\Element\FencedCode',
-            'Header'         => 'League\CommonMark\Block\Element\Header',
-            'HorizontalRule' => 'League\CommonMark\Block\Element\HorizontalRule',
+            'Heading'        => 'League\CommonMark\Block\Element\Heading',
             'HtmlBlock'      => 'League\CommonMark\Block\Element\HtmlBlock',
             'IndentedCode'   => 'League\CommonMark\Block\Element\IndentedCode',
             'ListBlock'      => 'League\CommonMark\Block\Element\ListBlock',
             'ListItem'       => 'League\CommonMark\Block\Element\ListItem',
             'Paragraph'      => 'League\CommonMark\Block\Element\Paragraph',
+            'ThematicBreak'  => 'League\CommonMark\Block\Element\ThematicBreak',
         ];
 
         foreach ($builtInClasses as $name => $fullyQualifiedName) {
@@ -183,14 +183,14 @@ class EnvironmentTest extends \PHPUnit_Framework_TestCase
         $mockRenderer = $this->getMock('League\CommonMark\Inline\Renderer\InlineRendererInterface');
 
         $builtInClasses = [
-            'Code'     => 'League\CommonMark\Inline\Element\Code',
-            'Emphasis' => 'League\CommonMark\Inline\Element\Emphasis',
-            'Html'     => 'League\CommonMark\Inline\Element\Html',
-            'Image'    => 'League\CommonMark\Inline\Element\Image',
-            'Link'     => 'League\CommonMark\Inline\Element\Link',
-            'Newline'  => 'League\CommonMark\Inline\Element\Newline',
-            'Strong'   => 'League\CommonMark\Inline\Element\Strong',
-            'Text'     => 'League\CommonMark\Inline\Element\Text',
+            'Code'       => 'League\CommonMark\Inline\Element\Code',
+            'Emphasis'   => 'League\CommonMark\Inline\Element\Emphasis',
+            'HtmlInline' => 'League\CommonMark\Inline\Element\HtmlInline',
+            'Image'      => 'League\CommonMark\Inline\Element\Image',
+            'Link'       => 'League\CommonMark\Inline\Element\Link',
+            'Newline'    => 'League\CommonMark\Inline\Element\Newline',
+            'Strong'     => 'League\CommonMark\Inline\Element\Strong',
+            'Text'       => 'League\CommonMark\Inline\Element\Text',
         ];
 
         foreach ($builtInClasses as $name => $fullyQualifiedName) {